refactor(skills): rename SkillCapability → AttachSkillCapability, SkillsDiscoveryCapability → SkillsCapability
AttachSkillCapability is now mount-only: it reconstructs SKILL.md from
DB fields and mounts to /.agents/skills/{name}/ so SkillsCapability
discovers it via VFS scan. No direct system prompt or tool contribution.

SkillsCapability (built-in "skills") remains the single source of
discovery, prompt injection, and activate_skill tooling.

Files renamed: skill.rs → attach_skill.rs, skills_discovery.rs → skills.rs
