+++ crates/core/src/lib.rs
@@ -159,10 +159,9 @@ pub use capabilities::{
     resolve_dependencies,
 };
 pub use capabilities::{
-    SKILL_CAPABILITY_PREFIX, SKILLS_DISCOVERY_CAPABILITY_ID, SKILLS_DISCOVERY_PATH,
-    SkillCapability, SkillInstructions, SkillMeta, SkillSource, SkillsDiscoveryCapability,
-    discover_skills_from_entries, is_skill_capability, parse_skill_capability_id,
-    skill_capability_id,
+    AttachSkillCapability, SKILL_CAPABILITY_PREFIX, SKILLS_CAPABILITY_ID, SKILLS_DISCOVERY_PATH,
+    SkillInstructions, SkillMeta, SkillSource, SkillsCapability, discover_skills_from_entries,
+    is_skill_capability, parse_skill_capability_id, skill_capability_id,
 };
 
 // Atoms re-exports (stateless atomic operations)
